requirements.txt
Pillow>=9.0.0
pytest>=6.2.4
